requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
colorclass==2.2.2
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
pip-upgrader==1.4.15
platformdirs==2.5.2
psycopg2==2.9.3
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==1.5
python-vuejs==1.0.9
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.3
terminaltables==3.1.10
tomli==2.0.1
tzdata==2022.4
unicodecsv==0.13.0
urllib3==1.26.12
whitenoise==6.2.0
yapf==0.32.0
